Huami 2021: Add missing cast to chunked type
Missed this file in the last commit.

+++ app/src/main/java/nodomain/freeyourgadget/gadgetbridge/service/devices/huami/Huami2021ChunkedDecoder.java
@@ -115,7 +115,7 @@ public class Huami2021ChunkedDecoder {
                 }
             }
             try {
-                huami2021Handler.handle2021Payload(currentType, buf);
+                huami2021Handler.handle2021Payload((short) currentType, buf);
             } catch (final Exception e) {
                 LOG.error("Failed to handle payload", e);
             }
